+++ db-queue-scheduler-spring/src/test/java/ru/yoomoney/tech/dbqueue/scheduler/BaseTest.java
@@ -1,16 +1,19 @@
 package ru.yoomoney.tech.dbqueue.scheduler;
 
 import org.junit.jupiter.api.BeforeAll;
-import org.postgresql.ds.PGSimpleDataSource;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.datasource.DataSourceTransactionManager;
-import org.springframework.transaction.support.TransactionTemplate;
+import org.testcontainers.containers.MSSQLServerContainer;
+import org.testcontainers.containers.OracleContainer;
 import org.testcontainers.containers.PostgreSQLContainer;
 import org.testcontainers.junit.jupiter.Container;
 import org.testcontainers.junit.jupiter.Testcontainers;
+import ru.yoomoney.tech.dbqueue.scheduler.db.DatabaseAccess;
+import ru.yoomoney.tech.dbqueue.scheduler.db.H2DatabaseAccessConfigurator;
+import ru.yoomoney.tech.dbqueue.scheduler.db.MsSqlDatabaseAccessConfigurator;
+import ru.yoomoney.tech.dbqueue.scheduler.db.OracleDatabaseAccessConfigurator;
+import ru.yoomoney.tech.dbqueue.scheduler.db.PostgresDatabaseAccessConfigurator;
 
-import javax.sql.DataSource;
 import java.util.concurrent.atomic.AtomicLong;
+import java.util.stream.Stream;
 
 /**
  * Common configuration
@@ -20,40 +23,32 @@ import java.util.concurrent.atomic.AtomicLong;
  */
 @Testcontainers
 public abstract class BaseTest {
-    private static final String SCHEDULED_TASKS_TABLE_DDL = "" +
-            "CREATE TABLE scheduled_tasks (" +
-            "  id                BIGSERIAL PRIMARY KEY," +
-            "  queue_name        TEXT NOT NULL," +
-            "  payload           TEXT," +
-            "  created_at        TIMESTAMP WITH TIME ZONE DEFAULT now()," +
-            "  next_process_at   TIMESTAMP WITH TIME ZONE DEFAULT now()," +
-            "  attempt           INTEGER                  DEFAULT 0," +
-            "  reenqueue_attempt INTEGER                  DEFAULT 0," +
-            "  total_attempt     INTEGER                  DEFAULT 0" +
-            ");" +
-            "CREATE UNIQUE INDEX scheduled_tasks_name_queue_name_uq ON scheduled_tasks (queue_name);";
+    @Container
+    private static final MSSQLServerContainer<?> mssqlServerContainer =
+            new MSSQLServerContainer<>("mcr.microsoft.com/mssql/server:2019-CU1-ubuntu-16.04").acceptLicense();
 
     @Container
     private static final PostgreSQLContainer<?> postgresqlContainer = new PostgreSQLContainer<>("postgres:9.5");
 
-    protected static JdbcTemplate jdbcTemplate;
-    protected static TransactionTemplate transactionTemplate;
+    @Container
+    private static final OracleContainer oracleContainer = new OracleContainer("gvenzl/oracle-xe:11.2.0.2-slim");
+
+    protected static DatabaseAccess postgres;
+    protected static DatabaseAccess oracle;
+    protected static DatabaseAccess mssql;
+    protected static DatabaseAccess h2;
+
     protected static AtomicLong uniqueCounter = new AtomicLong();
 
     @BeforeAll
-    public static void configurePostgres() {
-        DataSource dataSource = dataSource();
-        jdbcTemplate = new JdbcTemplate(dataSource);
-        transactionTemplate = new TransactionTemplate(new DataSourceTransactionManager(dataSource));
-
-        transactionTemplate.executeWithoutResult(status -> jdbcTemplate.update(SCHEDULED_TASKS_TABLE_DDL));
+    public static void configureDatabases() {
+        postgres = PostgresDatabaseAccessConfigurator.configure(postgresqlContainer);
+        oracle = OracleDatabaseAccessConfigurator.configure(oracleContainer);
+        mssql = MsSqlDatabaseAccessConfigurator.configure(mssqlServerContainer);
+        h2 = H2DatabaseAccessConfigurator.configure();
     }
 
-    private static DataSource dataSource() {
-        PGSimpleDataSource dataSource = new PGSimpleDataSource();
-        dataSource.setUrl(postgresqlContainer.getJdbcUrl());
-        dataSource.setPassword(postgresqlContainer.getPassword());
-        dataSource.setUser(postgresqlContainer.getUsername());
-        return dataSource;
+    public static Stream<DatabaseAccess> databaseAccessStream() {
+        return Stream.of(postgres, oracle, mssql, h2);
     }
 }